postgres_to_es/requirements.txt
psycopg2-binary==2.9
pydantic==1.10.5
elasticsearch==7.7.0
